more clean ups
