+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/pool_attributes.py
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 
-USER_AGENT_ORIGINAL = "user_agent.original"
+POOL_NAME = "pool.name"
 """
-Value of the [HTTP User-Agent](https://www.rfc-editor.org/rfc/rfc9110.html#field.user-agent) header sent by the client.
+Deprecated: Replaced by `db.client.connections.pool.name`.
 """
