+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/client_attributes.py
@@ -13,13 +13,6 @@
 # limitations under the License.
 
 
-
-
-
-
-
-
-
 CLIENT_ADDRESS = "client.address"
 """
 Client address - domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.
@@ -31,5 +24,3 @@ CLIENT_PORT = "client.port"
 Client port number.
 Note: When observed from the server side, and when communicating through an intermediary, `client.port` SHOULD represent the client port behind any intermediaries,  for example proxies, if it's available.
 """
-
-
