+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/error_attributes.py
@@ -13,12 +13,7 @@
 # limitations under the License.
 
 
-
-
-
-
 from enum import Enum
-
 
 ERROR_TYPE = "error.type"
 """
@@ -43,7 +38,6 @@ it's RECOMMENDED to:
 * Use a domain-specific attribute
 * Set `error.type` to capture all errors, regardless of whether they are defined within the domain-specific set or not.
 """
-
 
 
 class ErrorTypeValues(Enum):
