+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/exception_attributes.py
@@ -13,13 +13,6 @@
 # limitations under the License.
 
 
-
-
-
-
-
-
-
 EXCEPTION_ESCAPED = "exception.escaped"
 """
 SHOULD be set to true if the exception event is recorded at a point where it is known that the exception is escaping the scope of the span.
@@ -55,5 +48,3 @@ EXCEPTION_TYPE = "exception.type"
 """
 The type of the exception (its fully-qualified class name, if applicable). The dynamic type of the exception should be preferred over the static type in languages that support it.
 """
-
-
