+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/server_attributes.py
@@ -13,13 +13,6 @@
 # limitations under the License.
 
 
-
-
-
-
-
-
-
 SERVER_ADDRESS = "server.address"
 """
 Server domain name if available without reverse DNS lookup; otherwise, IP address or Unix domain socket name.
@@ -31,5 +24,3 @@ SERVER_PORT = "server.port"
 Server port number.
 Note: When observed from the client side, and when communicating through an intermediary, `server.port` SHOULD represent the server port behind any intermediaries, for example proxies, if it's available.
 """
-
-
