+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/service_attributes.py
@@ -13,13 +13,6 @@
 # limitations under the License.
 
 
-
-
-
-
-
-
-
 SERVICE_NAME = "service.name"
 """
 Logical name of the service.
@@ -30,5 +23,3 @@ SERVICE_VERSION = "service.version"
 """
 The version string of the service API or implementation. The format is not defined by these conventions.
 """
-
-
