+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/telemetry_attributes.py
@@ -13,12 +13,7 @@
 # limitations under the License.
 
 
-
-
-
-
 from enum import Enum
-
 
 TELEMETRY_SDK_LANGUAGE = "telemetry.sdk.language"
 """
@@ -40,7 +35,6 @@ TELEMETRY_SDK_VERSION = "telemetry.sdk.version"
 """
 The version string of the telemetry SDK.
 """
-
 
 
 class TelemetrySdkLanguageValues(Enum):
