+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/attributes/url_attributes.py
@@ -13,13 +13,6 @@
 # limitations under the License.
 
 
-
-
-
-
-
-
-
 URL_FRAGMENT = "url.fragment"
 """
 The [URI fragment](https://www.rfc-editor.org/rfc/rfc3986#section-3.5) component.
@@ -49,5 +42,3 @@ URL_SCHEME = "url.scheme"
 """
 The [URI scheme](https://www.rfc-editor.org/rfc/rfc3986#section-3.1) component identifying the used protocol.
 """
-
-
