+++ scripts/semconv/templates/registry/common.j2
@@ -27,12 +27,20 @@
 {% if str is defined and str is not none %}{{str}}{% else %}{{""}}{% endif %}
 {%- endmacro %}
 
+{%- macro remove_trailing_dots(str) -%}
+{%- if str[-1:] == '.' -%}{{ remove_trailing_dots(str[:-1]) }}{%- else -%}{{ str }}{%- endif -%}
+{%- endmacro -%}
+
 {%- macro to_docstring(str, prefix) -%}
-{{str | trim('.\n ') | comment_with_prefix(prefix)}}
+{{remove_trailing_dots(str | trim(' \n')) | comment_with_prefix(prefix)}}
 {%- endmacro %}
 
 {%- macro import_deprecated(semconv) -%}
-    {%- if (semconv | rejectattr("deprecated", "none") | list | count > 0) or (ctx.filter == "any" and semconv | selectattr("stability", "equalto", "stable") | list | count > 0) %}
+    {%- if (semconv | rejectattr("deprecated", "none") | list | count > 0) or (ctx.filter == "any" and semconv | selectattr("stability", "equalto", "stable") | list | count > 0) -%}
 from deprecated import deprecated
-    {%- endif %}
+    {%- endif -%}
 {%- endmacro-%}
+
+{%- macro print_value(type, value) -%}
+{%- if type == "string" -%}"{{value}}"{%-else-%}{{value}}{%-endif-%}
+{%- endmacro -%}
