+++ scripts/semconv/templates/registry/semantic_attributes.j2
@@ -20,7 +20,6 @@
 {% set attributes = ctx.attributes | sort(attribute="name") | list %}
 {% set enum_attributes = attributes | selectattr("type", "mapping") | list %}
 {% if enum_attributes | count > 0 %}from enum import Enum{% endif %}
-
 {{c.import_deprecated(enum_attributes)}}
 
 {%- macro attribute_name(attribute) -%}
@@ -65,7 +64,7 @@ class {{class_name}}(Enum):
     {%- for member in attribute.type.members %}
     {% set member_name = c.to_const_name(member.id) -%}
     {%- set doc_string=write_docstring(class_name + '.' + member_name, member.brief or member.id, "", member.deprecated, member.stability, "")-%}
-    {{member_name}} = "{{ member.value }}"
+    {{member_name}} = {{c.print_value(attribute.type | instantiated_type, member.value) }}
     {% if doc_string %}"""{{doc_string}}"""{% endif %}
     {%- endfor %}
 {% endfor %}
